[FLINK-23192][table] Move print sink and split config options
