+++ flink-table/flink-table-api-java-bridge/src/main/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -15,4 +15,4 @@
 
 org.apache.flink.connector.datagen.table.DataGenTableSourceFactory
 org.apache.flink.table.factories.BlackHoleTableSinkFactory
-org.apache.flink.table.factories.PrintTableSinkFactory
+org.apache.flink.connector.print.table.PrintTableSinkFactory
