+++ flink-table/flink-table-api-java-bridge/src/test/java/org/apache/flink/table/factories/PrintSinkFactoryTest.java
@@ -18,6 +18,8 @@
 
 package org.apache.flink.table.factories;
 
+import org.apache.flink.connector.print.table.PrintConnectorOptions;
+import org.apache.flink.connector.print.table.PrintTableSinkFactory;
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.catalog.Column;
 import org.apache.flink.table.catalog.ResolvedSchema;
@@ -29,8 +31,6 @@ import org.junit.Test;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.apache.flink.table.factories.PrintTableSinkFactory.PRINT_IDENTIFIER;
-import static org.apache.flink.table.factories.PrintTableSinkFactory.STANDARD_ERROR;
 import static org.apache.flink.table.factories.utils.FactoryMocks.createTableSink;
 
 /** Tests for {@link PrintTableSinkFactory}. */
@@ -46,8 +46,8 @@ public class PrintSinkFactoryTest {
     public void testPrint() {
         Map<String, String> properties = new HashMap<>();
         properties.put("connector", "print");
-        properties.put(PRINT_IDENTIFIER.key(), "my_print");
-        properties.put(STANDARD_ERROR.key(), "true");
+        properties.put(PrintConnectorOptions.PRINT_IDENTIFIER.key(), "my_print");
+        properties.put(PrintConnectorOptions.STANDARD_ERROR.key(), "true");
 
         DynamicTableSink sink = createTableSink(SCHEMA, properties);
         Assert.assertEquals("Print to System.err", sink.asSummaryString());
